Moved SuggestionIterator::SuggestionInternalData
Moved the declaration of SuggestionIterator::SuggestionInternalData
to suggestion_iterator.cpp. As a result, had to move the definition of
SuggestionResultSet::{begin,end}() there, too.
